allow switching of port-forward approaches when rootless and using slirp4netns
As of da7595a the default approach of providing port-forwarding in rootless mode has
switched and been hard-coded to rootlessport, which provides super performance.  The
downside/tradeoff of this switch is it takes away the ability to determine the source
of traffic being port-forwarded as that is a limitation of the new approach (see issue

This commit allows a user to specify if they want to revert to the previous approach
of leveraging slirp4netns add_hostfwd() api which although not as stellar performance
restores usefulness of seeing traffic origin.

The change should be transparent; when not specified, rootlessport will continue to be
used, however if specifying `--net slirp4netns:slirplisten` the old approach will be used.